changed store_fd type to uint8_t, this means we can have 255 files open consecutively, but the max size of the queue storing returned fds is only 255bytes at max

+++ src/libriccore/storage/storetypes.h
@@ -6,7 +6,7 @@
  * @brief Type of file descriptor
  * 
  */
-using store_fd = uint32_t;
+using store_fd = uint8_t;
  
 /**
  * @brief file mode values, copied directly from the sdfat impelmentation for ease,
